Pull out a bunch of things into submodules
