+++ tests/integrations/basic-fail/Cargo.stderr
@@ -501,7 +501,7 @@ Caused by:
 
 
 Location:
-    $DIR/src/lib.rs:LL:CC
+    $DIR/src/config.rs:LL:CC
 error: test failed, to rerun pass `--test ui_tests_invalid_program`
 
 Caused by:
